chore: move hooks out from utils
