+++ apps/wallet/src/features/Send/ConfirmTransaction/SenderInfo.tsx
@@ -2,7 +2,7 @@ import type { FC } from 'react';
 import { Image, StyleSheet } from 'react-native';
 import type { Networks } from '@walless/core';
 import { Text, View } from '@walless/gui';
-import { usePublicKeys } from 'utils/hooks';
+import { usePublicKeys } from 'hooks';
 import { getNetworkMetadata } from 'utils/transaction';
 import { useSnapshot } from 'valtio';
