+++ apps/wallet/src/features/Send/InputTransaction/CollectiblesTab.tsx
@@ -4,7 +4,7 @@ import type { AssetMetadata, Networks } from '@walless/core';
 import { Select, View } from '@walless/gui';
 import type { CollectibleDocument, CollectionDocument } from '@walless/store';
 import { NavButton } from 'components/NavButton';
-import { useNfts } from 'utils/hooks';
+import { useNfts } from 'hooks';
 import { useSnapshot } from 'valtio';
 
 import { txActions, txContext } from '../context';
