+++ apps/wallet/src/hooks/app.ts
