+++ apps/wallet/src/hooks/notification/index.web.ts
