+++ apps/wallet/src/hooks/wallet.ts
