+++ apps/wallet/src/hooks/widget.ts
