+++ apps/wallet/src/screens/Authentication/CreatePasscode/index.tsx
@@ -8,10 +8,10 @@ import {
 	TouchableOpacity,
 } from 'react-native';
 import { Passcode, Text, View } from '@walless/gui';
+import { useBiometricStatus, useSafeAreaInsets } from 'hooks';
 import { showError } from 'modals/Error';
 import { signInWithPasscode } from 'utils/auth';
 import { auth } from 'utils/firebase';
-import { useBiometricStatus, useSafeAreaInsets } from 'utils/hooks';
 import { hydrateEncryptionKey } from 'utils/native';
 import { navigate, ResetAnchors, resetRoute } from 'utils/navigation';
