+++ apps/wallet/src/stacks/Dashboard/ProfileIcon.tsx
@@ -2,8 +2,8 @@ import type { FC } from 'react';
 import type { ImageStyle } from 'react-native';
 import { Image } from 'react-native';
 import type { IconProps } from '@walless/icons';
+import { useSnapshot } from 'hooks';
 import { appState } from 'state';
-import { useSnapshot } from 'utils/hooks';
 
 export const ProfileIcon: FC<IconProps> = ({ size = 20 }) => {
 	const { profile } = useSnapshot(appState);
