+++ apps/wallet/tsconfig.json
@@ -30,6 +30,12 @@
       "state/*": [
         "./src/state/*"
       ],
+      "hooks": [
+        "./src/hooks"
+      ],
+      "hooks/*": [
+        "./src/hooks/*"
+      ],
       "components/*": [
         "./src/components/*"
       ],
